% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_more_plots.R
\name{compare_X_baseline_histogram}
\alias{compare_X_baseline_histogram}
\title{creates two-panel plot of histograms}
\usage{
compare_X_baseline_histogram(object, time, type, xmax = 20)
}
\arguments{
\item{object}{RPS_model object}

\item{time}{dbl, must be in object$time_vec}

\item{type}{string, one of "copula", "mixture"}
}
\value{

}
\description{
creates two-panel plot of histograms
}
