% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{compute_CVaR}
\alias{compute_CVaR}
\title{Compute CVaR under the P measure}
\usage{
compute_CVaR(time, kappa, q, c, dist)
}
\arguments{
\item{time}{dbl, time at which to compute CVaR}

\item{kappa}{dbl, compound Poisson parameter}

\item{q}{dbl, stressed VaR value}

\item{c}{dbl, VaR level, between 0 and 1}

\item{dist}{RPS_dist object}
}
\value{

}
\description{
Compute CVaR under the P measure
}
