% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{compute_VaR}
\alias{compute_VaR}
\title{computes VaR under the P measure using FST and stats::uniroot}
\usage{
compute_VaR(kappa, c, dist)
}
\arguments{
\item{kappa}{dbl, compound Poisson parameter}

\item{c}{dbl, level for VaR between 0 and 1}

\item{dist}{RPS_dist object}
}
\value{

}
\description{
computes VaR under the P measure using FST and stats::uniroot
}
