% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwise_correlation.R
\name{corr_CI}
\alias{corr_CI}
\title{compute correlation CI for Spearman's rho and Kendall's tau
based on Bonnett and Wright (2000)}
\usage{
corr_CI(r, n, method, alpha = 0.05)
}
\arguments{
\item{r}{correlation point estimate}

\item{n}{number of observations}

\item{method}{one of "spearman", "kendall"}

\item{alpha}{confidence level}
}
\value{

}
\description{
compute correlation CI for Spearman's rho and Kendall's tau
based on Bonnett and Wright (2000)
}
