% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_plots.R
\name{correlation_plot}
\alias{correlation_plot}
\title{correlation plot
plot correlation (Spearman's rho) over time}
\usage{
correlation_plot(object, time, type = "copula")
}
\arguments{
\item{object}{RPS_model object}

\item{time}{dbl, must be in object$time_vec}

\item{type}{string, one of "copula", "mixture"}
}
\value{

}
\description{
correlation plot
plot correlation (Spearman's rho) over time
}
