% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{CVaR_calc}
\alias{CVaR_calc}
\title{helper function for CVaR}
\usage{
CVaR_calc(X, alpha)
}
\arguments{
\item{X}{vector, data}

\item{alpha}{float, CVaR level}
}
\value{

}
\description{
helper function for CVaR
}
