% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{eta_CVaR}
\alias{eta_CVaR}
\title{Computes the optimal eta for CVaR stress}
\usage{
eta_CVaR(kappa, stress_parms, dist, endtime = 1, N = 8192)
}
\arguments{
\item{kappa}{dbl, compound Poisson parameter}

\item{stress_parms}{list, c = level of VaR & CVaR, q = stressed value of VaR,
s = stressed value of CVaR}

\item{dist}{RPS_dist object}

\item{endtime}{dbl}

\item{N}{int}
}
\value{

}
\description{
Computes the optimal eta for CVaR stress
}
