% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_parallel.R
\name{find_stress_parallel}
\alias{find_stress_parallel}
\title{Find optimal stress to VaR^P_alpha(X^1_{T/2}) to achieve a set stress to
VaR^Q_beta(X^1_T + X^2_T)}
\usage{
find_stress_parallel(
  floor,
  ceiling,
  perc_target,
  sum_VaR_P,
  alpha,
  beta,
  jump_dist,
  kappa = 5,
  dt = 0.01,
  tol = 1e-06,
  max_iter = 4
)
}
\arguments{
\item{floor}{dbl, floor of interval of stresses to search over, must be
below true stress}

\item{ceiling}{dbl, ceiling of interval of stresses to search over, must be
above true stress}

\item{perc_target}{target stress to VaR^Q_beta(X^1_T + X^2_T), as a percentage
i.e. 10, not 0.1}

\item{sum_VaR_P}{VaR_beta(X^1_T + X^2_T) under the baseline measure}

\item{alpha}{dbl, between 0 and 1 (strictly), alpha level for VaR at time 0.5}

\item{beta}{dbl, between 0 and 1 (strictly), alpha level for VaR at time 1}

\item{jump_dist}{RPS_dist_biv object}

\item{kappa}{dbl, baseline intensity}

\item{dt}{dbl, step size}

\item{tol}{dbl, tolerance at which to stop search}

\item{max_iter}{int, max number of iteration at which to stop search}
}
\value{
$table: tibble of stresses to VaR^P_alpha(X^1_{T/2}) and
resulting VaR^Q_beta(X^1_T + X^2_T), where T=1
$stress: optimal stress
}
\description{
Find optimal stress to VaR^P_alpha(X^1_{T/2}) to achieve a set stress to
VaR^Q_beta(X^1_T + X^2_T)
}
