% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST.R
\name{FST}
\alias{FST}
\title{Fourier Space Time-stepping (FST) algorithm}
\usage{
FST(
  x,
  t,
  dist,
  kappa,
  terminal_cond,
  grid_max,
  stress_time = 1,
  N = 5000,
  h = FALSE,
  y = NULL
)
}
\arguments{
\item{x}{vector, where to evaluate the FST}

\item{t}{double, time at which to evaluate the FST}

\item{dist}{RPS_dist object, contains jump size distribution}

\item{kappa}{double, compound Poisson parameter}

\item{terminal_cond}{double, end value}

\item{grid_max}{double, max grid value for FST}

\item{stress_time}{double, end time}

\item{N}{int, number of points in grid}

\item{h}{Boolean, function returns fraction of FST at x+y/FST at x if TRUE}

\item{y}{vector, optional argument for numerator if h = TRUE}
}
\value{
FST approximation of terminal_cond at x at time t
}
\description{
adapted from the Fourier Space Time-stepping (FST) library,
Copyright (C) 2007  Vladimir Surkov
https://sourceforge.net/projects/fst-framework/
license is GPL-3
}
