% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{h_CVaR}
\alias{h_CVaR}
\title{Computes h(t,x,y) for CVaR constraint}
\usage{
h_CVaR(x, y, t, kappa, dist, eta, stress_parms, N = 8192)
}
\arguments{
\item{x}{dbl}

\item{y}{dbl}

\item{t}{dbl}

\item{kappa}{dbl, compound Poisson parameter}

\item{dist}{RPS_dist objecrt}

\item{eta}{dbl}

\item{stress_parms}{list of parameters for stress; must contain:
q: stressed VaR value, time_stress: time at which stress occurs}

\item{N}{int}

\item{q}{dbl, stressed VaR value}
}
\value{

}
\description{
Computes h(t,x,y) for CVaR constraint
}
