% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST.R
\name{h_FST}
\alias{h_FST}
\title{Computes h(t,x,y)}
\usage{
h_FST(x, y, t, dist, eta, q, c, kappa, endtime = 1, N = 8192)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{t}{double}

\item{dist}{RPS_dist object, contains jump size distribution}

\item{eta}{double, optimal eta for VaR stress}

\item{q}{double, stressed VaR value}

\item{c}{double, VaR level to stress}

\item{kappa}{double, compound Poisson parameter}

\item{endtime}{double, end time}

\item{N}{int, number of points for FST method}
}
\value{

}
\description{
Computes h(t,x,y)
}
