% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.RPS_dist_univ}
\alias{is.RPS_dist_univ}
\title{Method for RPS_dist for is()}
\usage{
is.RPS_dist_univ(object)
}
\arguments{
\item{object}{RPS_dist}
}
\value{
Boolean
}
\description{
Method for RPS_dist for is()
}
