% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{mean.RPS_dist}
\alias{mean.RPS_dist}
\title{Method for RPS_dist for mean()}
\usage{
\method{mean}{RPS_dist}(object)
}
\arguments{
\item{object}{RPS_dist}
}
\value{
mean of distribution
}
\description{
Method for RPS_dist for mean()
}
