% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_RPS_dist}
\alias{new_RPS_dist}
\title{Defines RPS_dist class, class giving a distribution}
\usage{
new_RPS_dist(copula, margins, sim_fun, dens_fun, char_fun, mean_fun, parms)
}
\arguments{
\item{sim_fun}{function of x and parms, returns random draw generation function}

\item{dens_fun}{function of x and parms, returns density function}

\item{char_fun}{function of x and parms, returns characteristic function}

\item{mean_fun}{function of parms, returns mean}

\item{parms}{list, parameters for distribution}

\item{dist_fun}{function of x and parms, returns distribution function}
}
\value{

}
\description{
Defines RPS_dist class, class giving a distribution
}
