% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_RPS_model}
\alias{new_RPS_model}
\title{Defines RPS_model class}
\usage{
new_RPS_model(
  jump_dist,
  kappa,
  stress_type,
  stress_parms,
  paths,
  time_vec,
  kappa_Q
)
}
\arguments{
\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{kappa}{double, compound Poisson parameter}

\item{stress_type}{string, type of stress, one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list, parameters for stress}

\item{paths}{matrix, paths of stressed model}

\item{time_vec}{vector, times at which stressed model is evaluated}

\item{kappa_Q}{matrix, paths of stressed processed intensity}
}
\value{

}
\description{
Defines RPS_model class
}
