% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{P_expectation}
\alias{P_expectation}
\title{Computes P-expectation of e^(-eta2 * (x-q)) indicator(x >= q)
helper function for eta_CVaR and h_CVaR}
\usage{
P_expectation(x, t, kappa, q, eta2, dist, grid_max, time_stress = 1, N = 8192)
}
\arguments{
\item{x}{dbl}

\item{t}{dbl}

\item{kappa}{dbl, compound Poisson parameter}

\item{q}{dbl, stressed VaR value}

\item{eta2}{dbl}

\item{dist}{RPS_dist object}

\item{grid_max}{int}

\item{time_stress}{dbl}

\item{N}{int}
}
\value{

}
\description{
Computes P-expectation of e^(-eta2 * (x-q)) indicator(x >= q)
helper function for eta_CVaR and h_CVaR
}
