% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST.R
\name{P_prob}
\alias{P_prob}
\title{Helper function for compute_VaR
computes the P probability that X is less than q using FST}
\usage{
P_prob(kappa, q, dist)
}
\arguments{
\item{kappa}{dbl, compound Poisson parameter}

\item{q}{dbl}

\item{dist}{RPS_dist object}
}
\value{

}
\description{
Helper function for compute_VaR
computes the P probability that X is less than q using FST
}
