% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_plot.R
\name{path_plot}
\alias{path_plot}
\title{plot a RPS object
choose to plot paths only, or compare to baseline model,
or compare to kappa_Q}
\usage{
path_plot(
  object,
  Npaths,
  plot_type = "paths",
  quantiles = list(lower = 0.1, upper = 0.9)
)
}
\arguments{
\item{object}{RPS_model object}

\item{Npaths}{int, number of paths to plot}

\item{plot_type}{one of "compare_to_baseline" or "compare_to_kappa". If empty,
plots only stressed model}

\item{quantiles}{list containing lower and upper, specifies which quantiles
to highlight}
}
\value{

}
\description{
plot a RPS object
choose to plot paths only, or compare to baseline model,
or compare to kappa_Q
}
