% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_pathwise_correlation.R
\name{pathwise_corr}
\alias{pathwise_corr}
\title{Compute correlation with CI along a path}
\usage{
pathwise_corr(X1, X2, time_seq, method)
}
\arguments{
\item{X1}{vector 1}

\item{X2}{vector 2}

\item{time_seq}{times at which to evaluate vectors}

\item{method}{one of "pearson", "spearman", "kendall"}
}
\value{

}
\description{
Compute correlation with CI along a path
}
