% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_copula.R
\name{plot_copula}
\alias{plot_copula}
\title{Plot empirical copula under P and Q}
\usage{
plot_copula(object, time)
}
\arguments{
\item{object}{RPS_model object}

\item{time}{time at which to plot}
}
\value{

}
\description{
Plot empirical copula under P and Q
}
