% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_emp_copula.R
\name{plot_copula_contour}
\alias{plot_copula_contour}
\title{Plot empirical copula contour plot under P and Q}
\usage{
plot_copula_contour(object, time, type = "copula", show_true_ind = FALSE)
}
\arguments{
\item{object}{RPS_model object}

\item{time}{time at which to plot}

\item{type}{string, one of "copula", "mixture"}
}
\value{

}
\description{
Plot empirical copula contour plot under P and Q
}
