% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_plot_kappa_G.R
\name{plot_G_Q}
\alias{plot_G_Q}
\title{Plot dist of G^Q at multiple points in time and X}
\usage{
plot_G_Q(
  x_vec,
  t_vec,
  kappa,
  jump_dist,
  stress_type,
  stress_parms,
  N_out = 10000
)
}
\arguments{
\item{x_vec}{vector of state variable values}

\item{t_vec}{vector of dbl, between 0 and 1}

\item{kappa}{double, compound Poisson parameter}

\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{stress_type}{one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list of parameters for stress
VaR: q: stressed VaR value, c: VaR level, VaR_stress: multiplier for P-VaR
CVAR: q: stressed VaR value, c: VaR level, s: stressed CVaR level,
VaR_stress: multiplier for P-VaR, CVaR_stress: multiplier for P-CVaR}

\item{N_out}{int}
}
\value{

}
\description{
Plot dist of G^Q at multiple points in time and X
}
