% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_plot_kappa_G.R
\name{plot_kappa}
\alias{plot_kappa}
\title{Plot kappa^Q in 3D using plotly}
\usage{
plot_kappa(
  kappa,
  jump_dist,
  stress_type,
  stress_parms,
  end_time = 1,
  dt = 0.01
)
}
\arguments{
\item{kappa}{double, compound Poisson parameter}

\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{stress_type}{one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list of parameters for stress
VaR: q: stressed VaR value, c: VaR level, VaR_stress: multiplier for P-VaR
CVAR: q: stressed VaR value, c: VaR level, s: stressed CVaR level,
VaR_stress: multiplier for P-VaR, CVaR_stress: multiplier for P-CVaR}

\item{end_time}{float, when to end sim}

\item{dt}{float, step size in time}
}
\value{

}
\description{
Plot kappa^Q in 3D using plotly
}
