% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_plot.R
\name{plot_paths_biv}
\alias{plot_paths_biv}
\title{plot a RPS object
creates three-panel plot of paths under stressed model, paths under
baseline model, and kappa_Q}
\usage{
plot_paths_biv(
  object,
  Npaths,
  quantiles = list(lower = 0.1, upper = 0.9),
  indices = "random"
)
}
\arguments{
\item{object}{RPS_model object}

\item{Npaths}{int, number of paths to plot}

\item{quantiles}{list containing lower and upper, specifies which quantiles
to highlight}

\item{indices}{vector of indices to plot of length Npaths,
or ="random" to choose randomly}
}
\value{

}
\description{
plot a RPS object
creates three-panel plot of paths under stressed model, paths under
baseline model, and kappa_Q
}
