% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_plots.R
\name{plot_X_histogram}
\alias{plot_X_histogram}
\title{creates two-panel plot of histograms}
\usage{
plot_X_histogram(object, time)
}
\arguments{
\item{object}{RPS_model object}

\item{time}{dbl, must be in object$time_vec}
}
\value{

}
\description{
creates two-panel plot of histograms
}
