% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revpathsensitivity-package.R
\docType{package}
\name{revpathsensitivity-package}
\alias{revpathsensitivity}
\alias{revpathsensitivity-package}
\title{revpathsensitivity: Reverse path sensitivity for compound Poisson processes}
\description{
To be added.
}
\author{
\strong{Maintainer}: Emma Kroell \email{emma.kroell@mail.utoronto.ca} (\href{https://orcid.org/0000-0002-5813-0698}{ORCID})

}
\keyword{internal}
