% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{RPS_model}
\alias{RPS_model}
\title{Defines RPS_model class}
\usage{
RPS_model(
  model_type,
  jump_dist,
  kappa,
  stress_type,
  eta,
  stress_parms,
  paths,
  time_vec,
  intensity
)
}
\arguments{
\item{model_type}{string, one of "univariate", "bivariate_copula", or
"bivariate_mixture"}

\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{kappa}{double, compound Poisson parameter}

\item{stress_type}{string, type of stress, one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list, parameters for stress
VaR: q: stressed VaR value, c: VaR level, VaR_stress: multiplier for P-VaR
CVAR: q: stressed VaR value, c: VaR level, s: stressed CVaR level,
VaR_stress: multiplier for P-VaR, CVaR_stress: multiplier for P-CVaR
time_stress: time at which stress occurs; if NULL, assumed to be end_time}

\item{paths}{matrix, paths of stressed model}

\item{time_vec}{vector, times at which stressed model is evaluated}

\item{intensity}{matrix, paths of stressed processed intensity}
}
\value{

}
\description{
Defines RPS_model class
}
