% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_sim_model.R, R/2d_sim_model.R
\name{sim_baseline}
\alias{sim_baseline}
\title{function to get path of X under original probability measure}
\usage{
sim_baseline(object)

sim_baseline(object)
}
\arguments{
\item{object}{RPS_model object}
}
\value{
matrix, paths of original model

matrix, paths of original model
}
\description{
function to get path of X under original probability measure

function to get path of X under original probability measure
}
