% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_sim_model.R
\name{sim_baseline_biv}
\alias{sim_baseline_biv}
\title{#' function to get path of X under original probability measure
#'
#' @param object RPS_model object
#'
#' @return matrix, paths of original model
#' @export
#'
sim_baseline <- function(object){
with(object,{
# draw from the jump size distribution
withr::with_seed(720,{
Draws <- jump_dist$sim_fun(1e4,jump_dist$parms)}
\usage{
sim_baseline_biv(
  object = NULL,
  jump_dist = NULL,
  kappa = NULL,
  Npaths = NULL,
  time_vec = NULL
)
}
\description{
\preformatted{  Npaths <- dim(paths)[2]
  Nsteps <- length(time_vec)
}
}
\details{

}
