% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_model.R
\name{sim_G_kappa}
\alias{sim_G_kappa}
\title{compute G^Q and kappa^Q for Y ~ dist
helper function for stressed_sim()}
\usage{
sim_G_kappa(t, x, eta, kappa, q, c, dist, Draws, delta_y = 0.1, N = 10000)
}
\arguments{
\item{t}{A number,time}

\item{x}{A vector of numbers, state}

\item{eta}{float}

\item{kappa}{integer}

\item{q}{float, distorted quantile}

\item{c}{float, qantile to distort, c in (0,1)}

\item{dist}{RPS_dist object}

\item{Draws}{vector}

\item{delta_y}{float, step size for Riemann integration}

\item{N}{integer, number of draws from distribution}
}
\value{
A list containing:
\enumerate{
\item an vector of length(x), containing a draw from G^Q at for each x
\item a vector of length length(x) containing kappa^Q at each x
}
}
\description{
compute G^Q and kappa^Q for Y ~ dist
helper function for stressed_sim()
}
