% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_model_CVaR.R
\name{sim_G_kappa_CVaR}
\alias{sim_G_kappa_CVaR}
\title{compute G^Q and kappa^Q for Y ~ dist for CVaR
helper function for stressed_sim()}
\usage{
sim_G_kappa_CVaR(
  t,
  x,
  eta,
  kappa,
  stress_parms,
  dist,
  Draws,
  delta_y = 0.1,
  y_max = 100,
  tol = 1e-10
)
}
\arguments{
\item{t}{A number,time}

\item{x}{A vector of numbers, state}

\item{eta}{float}

\item{kappa}{integer}

\item{stress_parms}{list}

\item{dist}{RPS_dist object}

\item{Draws}{vector}

\item{delta_y}{float, step size for Riemann integration}

\item{y_max}{int, maximum for integration of h(t,x,y) along y-axes}

\item{tol}{dbl, cut off integration when prob less than tol}
}
\value{
A list containing:
\enumerate{
\item an vector of length(x), containing a draw from G^Q at for each x
\item a vector of length length(x) containing kappa^Q at each x
}
}
\description{
compute G^Q and kappa^Q for Y ~ dist for CVaR
helper function for stressed_sim()
}
