% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_model.R
\name{sim_G_kappa_mix}
\alias{sim_G_kappa_mix}
\title{compute G^Q and kappa^Q for Y ~ dist
helper function for stressed_sim()}
\usage{
sim_G_kappa_mix(
  t,
  x,
  eta,
  kappa,
  stress_type,
  stress_parms,
  dist,
  p,
  Draws_a,
  Draws_b,
  N_out = 1,
  delta_y = 0.1,
  y_max = 100,
  tol = 1e-10
)
}
\arguments{
\item{t}{A number,time}

\item{x}{A vector of numbers, state}

\item{eta}{float}

\item{kappa}{integer}

\item{stress_type}{one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list of parameters for stress
VaR: q: stressed VaR value, c: VaR level, VaR_stress: multiplier for P-VaR
CVAR: q: stressed VaR value, c: VaR level, s: stressed CVaR level,
VaR_stress: multiplier for P-VaR, CVaR_stress: multiplier for P-CVaR}

\item{dist}{RPS_dist object}

\item{N_out}{int, number of draws from distorted jump distribution}

\item{delta_y}{float, step size for Riemann integration}

\item{y_max}{int, maximum for integration of h(t,x,y) along y-axes}

\item{tol}{dbl, cut off integration when prob less than tol}

\item{Draws}{vector}
}
\value{
A list containing:
\enumerate{
\item an vector of length(x), containing a draw from G^Q at for each x
\item a vector of length length(x) containing kappa^Q at each x
}
}
\description{
compute G^Q and kappa^Q for Y ~ dist
helper function for stressed_sim()
}
