% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_sim_model.R
\name{stressed_sim}
\alias{stressed_sim}
\title{Simulate the stressed model}
\usage{
stressed_sim(
  kappa,
  jump_dist,
  stress_type = "VaR",
  stress_parms = list(c = c, q = q, s = s, VaR_stress = VaR_stress, CVaR_stress =
    CVaR_stress, time_stress = time_stress),
  X0 = 0,
  t0 = 0,
  end_time = 1,
  dt = 0.01,
  Npaths = 10000,
  beep = TRUE
)
}
\arguments{
\item{kappa}{double, compound Poisson parameter}

\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{stress_type}{one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list of parameters for stress
VaR: q: stressed VaR value, c: VaR level, VaR_stress: multiplier for P-VaR
CVAR: q: stressed VaR value, c: VaR level, s: stressed CVaR level,
VaR_stress: multiplier for P-VaR, CVaR_stress: multiplier for P-CVaR
time_stress: time at which stress occurs; if NULL, assumed to be end_time}

\item{X0}{float, initial condition for X}

\item{t0}{float, starting time}

\item{end_time}{float, when to end sim}

\item{dt}{float, step size in time}

\item{Npaths}{integer, number of paths}
}
\value{
RPS_model object
}
\description{
Simulate the stressed model
}
