% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Compute summary statistics at X_{t_delta} given X_{t0} = X0}
\usage{
summary_stats(
  X0,
  t0,
  t_delta,
  kappa,
  jump_dist,
  stress_type,
  stress_parms,
  end_time = 1,
  dt = 0.01,
  Npaths = 10000
)
}
\arguments{
\item{X0}{vector, initial conditions for X}

\item{t0}{float, starting time}

\item{t_delta}{float, time at which to calculate stats}

\item{kappa}{double, compound Poisson parameter}

\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{stress_type}{one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list of parameters for stress
VaR: q: stressed VaR value, c: VaR level, VaR_stress: multiplier for P-VaR
CVAR: q: stressed VaR value, c: VaR level, s: stressed CVaR level,
VaR_stress: multiplier for P-VaR, CVaR_stress: multiplier for P-CVaR
time_stress: time at which stress occurs; if NULL, assumed to be end_time}

\item{end_time}{float, when to end sim}

\item{dt}{float, step size in time}

\item{Npaths}{integer, number of paths}
}
\value{

}
\description{
Compute summary statistics at X_{t_delta} given X_{t0} = X0
}
