% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwise_correlation.R
\name{compare_correlation}
\alias{compare_correlation}
\title{Create plot comparing pathwise correlation under P and Q}
\usage{
compare_correlation(object, nsteps, method)
}
\arguments{
\item{object}{RPS_model object}

\item{nsteps}{number of steps at which to evaluate the correlation}

\item{method}{one of "pearson", "spearman", "kendall"}
}
\value{

}
\description{
Create plot comparing pathwise correlation under P and Q
}
