% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{eta2_eq}
\alias{eta2_eq}
\title{Equation to be solved to find eta_2}
\usage{
eta2_eq(eta, dist, q, s, kappa, grid_max, endtime = 1, N = 8192)
}
\arguments{
\item{eta}{dbl}

\item{dist}{RPS_dist object}

\item{q}{dbl, stressed VaR value}

\item{s}{dbl, stressed CVaR value}

\item{kappa}{dbl, compound Poisson parameter}

\item{N}{}
}
\value{

}
\description{
Equation to be solved to find eta_2
}
