% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{h_CVaR}
\alias{h_CVaR}
\title{Computes h(t,x,y) for CVaR constraint}
\usage{
h_CVaR(x, y, t, kappa, q, dist, eta, endtime = 1, N = 8192)
}
\arguments{
\item{x}{dbl}

\item{y}{dbl}

\item{t}{dbl}

\item{kappa}{dbl, compound Poisson parameter}

\item{q}{dbl, stressed VaR value}

\item{dist}{RPS_dist objecrt}

\item{eta}{dbl}

\item{endtime}{dbl}

\item{N}{int}
}
\value{

}
\description{
Computes h(t,x,y) for CVaR constraint
}
