% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{h_VaR}
\alias{h_VaR}
\title{Computes h(t,x,y) for VaR constraint}
\usage{
h_VaR(x, y, t, dist, eta, kappa, stress_parms, N = 8192)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{t}{double}

\item{dist}{RPS_dist object, contains jump size distribution}

\item{eta}{double, optimal eta for VaR stress}

\item{kappa}{double, compound Poisson parameter}

\item{stress_parms}{list of parameters for stress; must contain:
q: stressed VaR value, time_stress: time at which stress occurs}

\item{N}{int, number of points for FST method}

\item{q}{double, stressed VaR value}
}
\value{

}
\description{
Computes h(t,x,y) for VaR constraint
}
