% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FST_implementations.R
\name{P_prob}
\alias{P_prob}
\title{Computes the P probability that X is less than q using FST}
\usage{
P_prob(x, t, kappa, q, dist, grid_max = 50, time_stress = 1, N = 8192)
}
\arguments{
\item{x}{dbl}

\item{t}{dbl}

\item{kappa}{dbl, compound Poisson parameter}

\item{q}{dbl}

\item{dist}{RPS_dist object}

\item{grid_max}{int}

\item{time_stress}{dbl}

\item{N}{int}
}
\value{

}
\description{
Computes the P probability that X is less than q using FST
}
