% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_plot.R
\name{path_plot}
\alias{path_plot}
\title{plot a RPS object
creates three-panel plot of paths underr stressed model, paths under
baseline model, and kappa_Q}
\usage{
path_plot(
  object,
  Npaths,
  quantiles = list(lower = 0.1, upper = 0.9),
  indices = "random"
)
}
\arguments{
\item{object}{RPS_model object}

\item{Npaths}{int, number of paths to plot}

\item{quantiles}{list containing lower and upper, specifies which quantiles
to highlight}

\item{indices}{vector of indices to plot of length Npaths,
or ="random" to choose randomly}
}
\value{

}
\description{
plot a RPS object
creates three-panel plot of paths underr stressed model, paths under
baseline model, and kappa_Q
}
