% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_emp_copula.R
\name{plot_copula}
\alias{plot_copula}
\title{Plot empirical copula under P and Q}
\usage{
plot_copula(object, time, type = "copula")
}
\arguments{
\item{object}{RPS_model object}

\item{time}{time at which to plot}

\item{type}{string, one of "copula", "mixture"}
}
\value{

}
\description{
Plot empirical copula under P and Q
}
