% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_model.R
\name{stressed_sim}
\alias{stressed_sim}
\title{Simulate the stressed model}
\usage{
stressed_sim(
  kappa,
  jump_dist,
  stress_type = "VaR",
  stress_parms = list(c = c, q = q),
  Npaths = 10000,
  endtime = 1,
  dt = 0.01
)
}
\arguments{
\item{kappa}{double, compound Poisson parameter}

\item{jump_dist}{RPS_dist object, contains jump size distribution}

\item{stress_type}{one of "VaR", "VaR & CVaR"}

\item{stress_parms}{list of parameters for stress
VaR: q: stressed VaR value, c: VaR level
CVAR: TBA}

\item{Npaths}{integer, number of paths}

\item{endtime}{float, when to end sim}

\item{dt}{float, step size in time}
}
\value{
RPS_model object
}
\description{
Simulate the stressed model
}
